refactor: relocated auth and register window in common folder
