+++ client/src/components/controller/index/register-controller.ts
@@ -2,7 +2,7 @@ import INotify from "../../interfaces/i-notify";
 import UserModel from "../../models/index/data-model/user-model";
 import RegisterModel from "../../models/index/state-model/register-model";
 import AuthView from "../../view/index/auth/auth-view";
-import RegisterWindowView from "../../view/index/register-window/register-window-view";
+import RegisterWindowView from "../../view/common/register-window/register-window-view";
 import View from "../../view/index/view";
 import { AppEvents } from "../app-events";
