+++ client/src/components/view/common/register-window/register-window-view.ts
@@ -1,10 +1,10 @@
 import { AppEvents } from "../../../controller/app-events";
 import Observer from "../../../controller/observer";
 import INotify from "../../../interfaces/i-notify";
-import View from "../view";
-import './auth-window.scss';
-import '../../common/button.scss';
-export default class AuthWindowView extends View implements INotify {
+import View from "../../index/view";
+import './register-window.scss';
+
+export default class RegisterWindowView extends View implements INotify {
     private readonly TAG_CONTAINER = 'section';
     private readonly TAG_WINDOW = 'form';
     private readonly TAG_HEADER = 'h5';
@@ -12,19 +12,20 @@ export default class AuthWindowView extends View implements INotify {
     private readonly TAG_BUTTON = 'button';
     private readonly TAG_LABEL = 'label';
     private readonly TAG_FIELD = 'input';
-    private readonly CLASS_CONTAINER = 'window-auth__wrapper';
-    private readonly CLASS_WINDOW = 'window-auth';
-    private readonly CLASS_HEADER = 'window-auth__header';
-    private readonly CLASS_ROW_CONTAINER = 'window-auth__row-container';
-    private readonly CLASS_LABEL = 'window-auth__info';
-    private readonly CLASS_FIELD = 'window-auth__field';
-    private readonly CLASS_FIELD_INVALID = 'window-auth__field_invalid';
+    private readonly CLASS_CONTAINER = 'window-register__wrapper';
+    private readonly CLASS_WINDOW = 'window-register';
+    private readonly CLASS_HEADER = 'window-register__header';
+    private readonly CLASS_ROW_CONTAINER = 'window-register__row-container';
+    private readonly CLASS_LABEL = 'window-register__info';
+    private readonly CLASS_FIELD = 'window-register__field';
+    private readonly CLASS_FIELD_INVALID = 'window-register__field_invalid';
     private readonly CLASS_BUTTON = 'big__button';
 
-    private readonly TEXT_HEADER = 'User authorization';//TODO (local) выносится в локализацию
+    private readonly TEXT_HEADER = 'Registration new user';//TODO (local) выносится в локализацию
     private readonly TEXT_FIELD_LOGIN = 'Login';//TODO (local) выносится в локализацию
     private readonly TEXT_FIELD_PASS = 'Password';//TODO (local) выносится в локализацию
-    private readonly TEXT_BUTTON_LOGIN = 'Login user';//TODO (local) выносится в локализацию
+    private readonly TEXT_FIELD_EMAIL = 'Email';//TODO (local) выносится в локализацию
+    private readonly TEXT_BUTTON_LOGIN = 'Register';//TODO (local) выносится в локализацию
     private readonly TEXT_BUTTON_CANCEL = 'Cancel';//TODO (local) выносится в локализацию
     private readonly ID_FIELD_LOGIN = 'login';
     private readonly ID_FIELD_PASS = 'password';
@@ -36,14 +37,14 @@ export default class AuthWindowView extends View implements INotify {
     constructor(observer: Observer) {
         super(observer);
         this.createWindowElement();
-        this._observer.addSender(AppEvents.AUTH_CLICK_BUTTON, this);
+        this._observer.addSender(AppEvents.REGISTER_CLICK_BUTTON, this);
     }
     getCurrentElement(): HTMLElement {
         return this._windowElement;
     }
     notify(nameEvent: AppEvents): AppEvents | void {
-        if(nameEvent === AppEvents.AUTH_CLICK_BUTTON) {
-            this._observer.notify(AppEvents.AUTH_SHOW_WINDOW, this);  
+        if(nameEvent === AppEvents.REGISTER_CLICK_BUTTON) {
+            this._observer.notify(AppEvents.REGISTER_SHOW_WINDOW, this);  
         }
     }
     setWindowVisibilityState(state: boolean): void {
@@ -53,21 +54,26 @@ export default class AuthWindowView extends View implements INotify {
             this._windowElement.style.visibility = 'hidden';
         }
     }
-    successLogInHandler() {
+    successRegistrationHandler() {
         this.setWindowVisibilityState(false);//TODO убрать прямое изменение состояния
-        this._observer.notify(AppEvents.AUTH_LOGIN_USER_SUCCESS, this);
+        this._observer.notify(AppEvents.REGISTER_USER_SUCCESS, this);
     }
-    failLogInHandler(result: Map<string, string>) {
+    failRegistrationHandler(result: Map<string, string>) {
         //TODO запилить отдельный компонент инфо окна для всех страниц
         alert(result.get('message'));
     }
-    private logInUser() {
+    private closeWindow() {
+        event?.preventDefault();
+        this._observer.notify(AppEvents.REGISTER_HIDE_WINDOW, this);
+    }
+    private registerUser() {
         event?.preventDefault();
         if(this.checkFormData()) {
             let params = new Map<string, string>();
             params.set(this.ID_FIELD_LOGIN, this._formElement.login.value);
+            params.set(this.ID_FIELD_EMAIL, this._formElement.email.value);
             params.set(this.ID_FIELD_PASS, this._formElement.password.value);
-            this._observer.notify(AppEvents.AUTH_LOGIN_USER, this, params);
+            this._observer.notify(AppEvents.REGISTER_USER, this, params);
         }
     }
     private checkFormData(): boolean {
@@ -84,11 +90,13 @@ export default class AuthWindowView extends View implements INotify {
         } else {
             this._formElement.password.classList.remove(this.CLASS_FIELD_INVALID);
         }
+        if (this._formElement.email.value == '') {
+            this._formElement.email.classList.add(this.CLASS_FIELD_INVALID);
+            result = false;
+        } else {
+            this._formElement.email.classList.remove(this.CLASS_FIELD_INVALID);
+        }
         return result;
-    }
-    private closeWindow() {
-        event?.preventDefault();
-        this._observer.notify(AppEvents.AUTH_HIDE_WINDOW, this);
     }
     private createWindowElement(): void {
         this._windowElement.classList.add(this.CLASS_CONTAINER);
@@ -113,6 +121,18 @@ export default class AuthWindowView extends View implements INotify {
 
         rowElement = document.createElement(this.TAG_ROW_CONTAINER);
         rowElement.classList.add(this.CLASS_ROW_CONTAINER);
+        const textEmail = document.createElement(this.TAG_LABEL);
+        textEmail.classList.add(this.CLASS_LABEL);
+        textEmail.textContent = this.TEXT_FIELD_EMAIL;
+        const fieldEmail = document.createElement(this.TAG_FIELD);
+        fieldEmail.classList.add(this.CLASS_FIELD);
+        fieldEmail.setAttribute('id', this.ID_FIELD_EMAIL);
+        rowElement.insertAdjacentElement('beforeend', textEmail);
+        rowElement.insertAdjacentElement('beforeend', fieldEmail);
+        this._formElement.insertAdjacentElement('beforeend', rowElement);
+
+        rowElement = document.createElement(this.TAG_ROW_CONTAINER);
+        rowElement.classList.add(this.CLASS_ROW_CONTAINER);
         const textPassword = document.createElement(this.TAG_LABEL);
         textPassword.classList.add(this.CLASS_LABEL);
         textPassword.textContent = this.TEXT_FIELD_PASS;
@@ -129,7 +149,9 @@ export default class AuthWindowView extends View implements INotify {
         const loginButton = document.createElement(this.TAG_BUTTON);
         loginButton.classList.add(this.CLASS_BUTTON);
         loginButton.textContent = this.TEXT_BUTTON_LOGIN;
-        loginButton.addEventListener('click', this.logInUser.bind(this));
+        loginButton.setAttribute('type', 'submit');
+
+        loginButton.addEventListener('click', this.registerUser.bind(this));
         const cancelButton = document.createElement(this.TAG_BUTTON);
         cancelButton.classList.add(this.CLASS_BUTTON);
         cancelButton.textContent = this.TEXT_BUTTON_CANCEL;
