+++ client/src/components/view/common/register-window/register-window.scss
@@ -1,4 +1,4 @@
-.window-auth__wrapper {
+.window-register__wrapper {
     visibility: hidden;
 
     position: absolute;
@@ -7,12 +7,12 @@
     right   : 0;
 
     height: 100vh;
-    width : 100vw;
+    // width : 100vw;
 
     background: $modal_background-color;
 }
 
-.window-auth {
+.window-register {
     position: absolute;
     left    : 35%;
     top     : 40%;
@@ -23,18 +23,18 @@
     align-items    : center;
 
     height    : 30%;
-    min-height: 200px;
+    min-height: 260px;
     width     : 30%;
 
     background-color: $window_background-color;
 }
 
-.window-auth__header {
+.window-register__header {
     font-size: 1rem;
     margin   : 1rem;
 }
 
-.window-auth__row-container {
+.window-register__row-container {
     display        : flex;
     flex-direction : row;
     align-items    : center;
@@ -46,12 +46,12 @@
     padding: 0% 15%;
 }
 
-.window-auth__info {
+.window-register__info {
     width: 20%;
     text-align: end;
 }
 
-.window-auth__field {
+.window-register__field {
     width: 60%;
 
     font-size    : 16px;
@@ -62,6 +62,6 @@
     cursor       : pointer;
 }
 
-.window-auth__field_invalid {
+.window-register__field_invalid {
     border       : 1px solid $border-color-error;
 }
